Add an option to print a short summary.
With the -s/--short argument, each change is printed in a single line. New tasks
are still printed in the multi-line format.

Field names are now colored in magenta.

The crate `anyhow` is added to improve some error messages.

The change `clap` is added to make easier to parse arguments from the prompt.